requirements.txt
# ── Core API ────────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.2.0

# ── Scheduling ───────────────────────────────────────────────────────────────
apscheduler>=3.10.4

# ── Data Collection ──────────────────────────────────────────────────────────
pynput>=1.7.6
pillow>=10.3.0
pywin32>=306
opencv-python>=4.9.0.80
sounddevice>=0.4.6
scipy>=1.13.0
watchdog>=4.0.0
imagehash>=4.3.1
psutil>=5.9.8
requests>=2.31.0

# ── OCR ──────────────────────────────────────────────────────────────────────
easyocr>=1.7.1

# ── Embeddings & ML ──────────────────────────────────────────────────────────
sentence-transformers>=3.0.0
open-clip-torch>=2.24.0
torch>=2.3.0
torchvision>=0.18.0

# ── Reranking ────────────────────────────────────────────────────────────────
# cross-encoder is included in sentence-transformers

# ── Vector Database ───────────────────────────────────────────────────────────
chromadb>=0.5.0

# ── Config ───────────────────────────────────────────────────────────────────
pyyaml>=6.0.1
python-dotenv>=1.0.0

# ── Logging ──────────────────────────────────────────────────────────────────
loguru>=0.7.2

# ── System Tray ──────────────────────────────────────────────────────────────
pystray>=0.19.5

# ── Document Parsing ─────────────────────────────────────────────────────────
python-docx>=1.1.0
pypdf>=4.2.0

# ── Cryptography (optional at-rest encryption) ────────────────────────────────
cryptography>=42.0.0

# ── NLP / NER (privacy pipeline + tagging) ───────────────────────────────────
spacy>=3.7.0

# ── Frontier API clients (intelligence pipeline) ─────────────────────────────
anthropic>=0.25.0
openai>=1.30.0

# ── MCP server ───────────────────────────────────────────────────────────────
mcp>=1.0.0

# ── Original keylog.py deps (preserved) ──────────────────────────────────────
pyaudio>=0.2.14
